refactor(admin): use page-level tabs for alerting instead of sidebar submenu
Replace the collapsible sidebar submenu with a single 'Alerting' sidebar
entry that opens a tabbed page with Error Rate, TTFB, and Model Status
tabs. Follows the existing admin tab pattern (URL-synced ?tab= param,
underline-style triggers).

- Extract AlertingContent and TtfbAlertingContent into standalone components
- Redirect /admin/alerting-ttfb and /admin/model-status to the tabbed page
- Model Status tab shows 'Under construction' placeholder
